Dev relationship

See merge request bachelor/2020/group-107/bachelor!99

+++ bachelor/sim/src/entity/components/tags.h
@@ -41,9 +41,9 @@ inline std::string tag_to_string(ETag tag)
                                        "Gather", "Delete", "Reserved",  "Carnivore", "Herbivore", "Omnivore",
                                        "Meat",   "Veggie", "Inventory", "Consume"};
 
-    for (unsigned i = 0u; i < tag_names.size(); i++)
+    for (uint64_t i = 0ULL; i < tag_names.size(); i++)
     {
-        if (tag & 1 << i)
+        if (tag & 1ULL << i)
         {
             result += " " + tag_names[i];
         }
